pxqmr/pyproject.toml
[project]
name = "pxqmr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fuzzywuzzy>=0.18.0",
    "mcp[cli]>=1.6.0",
    "python-levenshtein>=0.27.1",
    "python-reapy>=0.10.0",
]

[tool.uv.workspace]
members = [
    "heelo.py",
]
